+++ src/Symfony/Component/HttpKernel/Controller/ArgumentResolver.php
@@ -60,7 +60,7 @@ final class ArgumentResolver implements ArgumentResolverInterface
                     if ($attribute->disabled) {
                         $disabledResolvers[$attribute->resolver] = true;
                     } elseif ($resolverName) {
-                        throw new \LogicException(sprintf('You can only pin one resolver per argument, but argument "$%s" of "%s()" has more.', $metadata->getName(), $this->getPrettyName($controller)));
+                        throw new \LogicException(sprintf('You can only pin one resolver per argument, but argument "$%s" of "%s()" has more.', $metadata->getName(), $metadata->getControllerName()));
                     } else {
                         $resolverName = $attribute->resolver;
                     }
@@ -118,7 +118,7 @@ final class ArgumentResolver implements ArgumentResolverInterface
                 }
             }
 
-            throw new \RuntimeException(sprintf('Controller "%s" requires the "$%s" argument that could not be resolved. '.($reasonCounter > 1 ? 'Possible reasons: ' : '').'%s', $this->getPrettyName($controller), $metadata->getName(), implode(' ', $reasons)));
+            throw new \RuntimeException(sprintf('Controller "%s" requires the "$%s" argument that could not be resolved. '.($reasonCounter > 1 ? 'Possible reasons: ' : '').'%s', $metadata->getControllerName(), $metadata->getName(), implode(' ', $reasons)));
         }
 
         return $arguments;
@@ -136,22 +136,5 @@ final class ArgumentResolver implements ArgumentResolverInterface
             new DefaultValueResolver(),
             new VariadicValueResolver(),
         ];
-    }
-
-    private function getPrettyName($controller): string
-    {
-        if (\is_array($controller)) {
-            if (\is_object($controller[0])) {
-                $controller[0] = get_debug_type($controller[0]);
-            }
-
-            return $controller[0].'::'.$controller[1];
-        }
-
-        if (\is_object($controller)) {
-            return get_debug_type($controller);
-        }
-
-        return $controller;
     }
 }