+++ src/Symfony/Component/HttpKernel/ControllerMetadata/ArgumentMetadata.php
@@ -31,6 +31,7 @@ class ArgumentMetadata
         private mixed $defaultValue,
         private bool $isNullable = false,
         private array $attributes = [],
+        private string $controllerName = 'n/a',
     ) {
         $this->isNullable = $isNullable || null === $type || ($hasDefaultValue && null === $defaultValue);
     }
@@ -134,5 +135,10 @@ class ArgumentMetadata
         }
 
         return $attributes;
+    }
+
+    public function getControllerName(): string
+    {
+        return $this->controllerName;
     }
 }