+++ src/Symfony/Bridge/Doctrine/CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
  * Allow `EntityValueResolver` to return a list of entities
  * Add support for auto-closing idle connections
  * Allow validating every class against `UniqueEntity` constraint
+ * Deprecate auto-mapping of entities in favor of mapped route parameters
 
 7.0
 ---